feat: update training configuration and model handling
- Increase learning rate from 1e-4 to 1e-1 in prompt_learning.yaml
- Remove dummy text ids and fix attention mask handling in pl_handle.py
- Switch precision from 16-mixed to bf16-mixed in train scripts
- Update test configuration to match training changes

+++ tests/test_train_pl.py
@@ -62,7 +62,7 @@ def train(cfg: DictConfig) -> None:
         gradient_clip_val=1.0,  # NOTE: gradient clipping will be normed
         # gradient_clip_algorithm="value",
         sync_batchnorm=True,
-        precision="16-mixed",
+        precision="bf16-mixed",
         logger=None,
         # WARN: will slow down the training process, just for debug now
         # detect_anomaly=True,
